fix bug of set extension typing
When adding a set extension expression, we need to know the data type of the elements inside (to be used in CppTranslation for use of template). When inside a set extension expression, using elements[index].getType() when the element inside the set extension is a freeIdentifier returns null. If this happens, we cant then turn it to a string and feed it to the getDataType function of the visitor. IN THE CASE OF A FREEIDENTIFIER, it will use a Map that the CppAST has of every freeIdentifier defined in contexts and machines and its respective DataType. IN THE CASE the element inside the SetExtension being used to determine the data type isnt a freeIdentifier (its something like a number), WE CAN use the .getType().toString() and feed it to the getDataType and get the datatype back, just like when determining the type of constants and variables.